+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.2.0
+ - moved onto Go modules;
+ - `stopwords#registerStopWords` made private.
+
 ## 0.1.0
  - first release, and it will have stop-words registered.